refactor(breeze): move all Breeze-specific code under src/Breeze
Everything that knows about the Breeze plugin now lives under src/Breeze/
(namespace Parisek\TimberKit\Breeze), so a project that does not run Breeze
can see in one directory what is dead weight.

A new architecture test enforces the boundary rather than documenting it:
naming Breeze anywhere else under src/ fails the build. StarterBase is the
sole documented exception, because it keeps the opt-in flags and those flag
names are public API — six projects in the fleet already set them.

BreezeWarmupSitemap becomes Breeze\WarmupSitemap. The old class name keeps
resolving through a class_alias in compat/aliases.php, wired as a composer
files autoload entry; the class is final, which rules out a subclass shim.
Dropping the alias outright was rejected: the old name has shipped, so
removing it would break any consumer still referencing it.

No behaviour changes. The warmup logic is untouched — this is a relocation
plus the compatibility shim.
